Resolves #2661 

This should give a decent speedup for loading Gallery.

Before this PR, we would go one label type at a time, and check each label of that label type for valid GSV imagery in parallel. Now instead of going one label type at a time, we combine all the label types together. This requires some more complicated accounting, keeping track of a mapping from each label type to a bunch of extra metadata. In the end, it only adds  a few more lines of code though.

Ultimately, we found around a 40% speedup in the portion of the page load that involves checking for GSV imagery. I tested on a database dump with 186k labels, testing 10 page loads in each case. In the situation where the whole database was already up-to-date on which labels had imagery, the time went from 1.75 seconds to 1.06 seconds (39% speedup). And in the situation where all labels are marked as not-expired (and we are therefore going to hit a lot of instances where imagery actually _is_ expired), we went from 5.74 seconds to 3.27 seconds (43% speedup).

In addition, I parameterized some of the code for checking for GSV imagery since we have 5 functions with this duplicated code, and our unwieldy `LabelTable.scala` file is over 1,000 lines long right now :)

##### Things to check before submitting the PR <!-- if something doesn't apply, just check the box or remove the line -->
- [x] I've written a descriptive PR title.
- [x] I've added/updated comments for large or confusing blocks of code.